refactor: rename the camelCase import paths in tests and lint config
