+++ frontend/eslint.config.js
@@ -17,8 +17,8 @@ const restrictedApiIndexImports = [
   '@/api/institutions',
   '@/api/merchants',
   '@/api/tags',
-  '@/api/taxAdvantagedCategories',
-  '@/api/transactionImports',
+  '@/api/tax-advantaged-categories',
+  '@/api/transaction-imports',
   '@/api/transactions',
   '@/api/user',
   '@/api/version',
